Projectile
Added a check to ensure that a player can't spam projectiles by tapping
the key repeatedly.

I'll say projectile is complete - we can use this as a class for other
projectiles in the game, and we can always add sound effects and other
functionality later as the game sees fit.