Update movement / SoundFX
Updated movement so it depends on time and not on framerate.. therefore
we can adjust the framerate of the game and nothing should change
(except how smooth it looks). Also, when a sprite collides, it should be
as close to the solid as possible, which helps a bit in squeezing into
tight spots.

-Added FreeMovingSprite class to encapsulate movement
-Adjusted child classes (Character and Projectile) accordingly.
-Added SoundFX class to hold all sound effects (and it's easier to mute
sounds that way)