Added clue mechanic
The clue mechanic is a bit simple at the moment, but it's functional. To
use it, go to Clues in the pause menu. You can select one or more clues
by pressing the action/confirm key (there's currently no indicator for
whether a clue is selected or not; I might bring this up in the
meeting). Then you can press Right to link the clues together and see
what happens (not the best solution; I'll probably bring this up as
well).