Maps
Added maps according to our blueprint.

Maps that are missing: Computer Lab, Student Lounge, Parking Lot, Outer
Entrance, Principal's Office, Gym.

Besides that the rest are done/I'll add more detail as we go along and
get new tilesets and I'll be working on the other maps as well. The maps
are still not implemented to the game though.