Organization!!1!
Placed content and source into folders for organization. Some of the
categories are a bit arbitrary, so they can be reordered/changed for
clarity.. I think it's mostly good, though.

--Source is now in src folder, which is split into Folders representing
different types of source. Game objects go in "Objects", menu panels go
in "Panels", etc. When adding new source to these folders, make sure the
namespace is always "SunsetHigh"

-- Content folder now has subdirectories for each type of content. These
folders should be self-explanatory - note that Character sprites go in
"Characters", and all others go in "Sprites" (and "CharactersTemp" is
for sprites we didn't make)

-- !!IMPORTANT: Added Directories.cs static class, which has string
literals for the source to reference content. Prepend the appropriate
directory name to the name of the content file when using it.